move integration test to test/test-application
