+++ .circleci/config.yml
@@ -24,7 +24,7 @@ jobs:
             cp -r typescript/test/protos ./.baseline-test-out
             cd .baseline-test-out 
             npm install
-            cd ../typescript/test_application
+            cd ../typescript/test/test_application
             npm install
             npm test
       - run:
