+++ typescript/test/test_application/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "test-application",
+  "version": "0.0.1",
+  "description": "Test application for generated client library.",
+  "engines": {
+    "node": ">=8.13.0"
+  },
+  "scripts": {
+    "test": "bash showcase-setup.sh"
+  },
+  "author": "Google LLC",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "showcase": "file:///../../../.baseline-test-out",
+    "@grpc/grpc-js": "^0.5.2",
+    "mocha": "^6.2.0"
+  }
+}
